third_party/blink/web_tests/external/wpt/scroll-animations/css/view-timeline-pseudo-on-scroller-expected.txt

This is a testharness.js-based test.
[FAIL] view() on pseudo-element attaches to parent scroll container
  assert_equals: expected "rgb(0, 0, 255)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
